# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
      "AccentPhrase": {
        "description": "アクセント句ごとの情報。",
        "properties": {
          "accent": {
            "description": "アクセント箇所",
            "minimum": 1.0,
            "title": "Accent",
            "type": "integer"
          },
          "is_interrogative": {
            "default": false,
            "description": "疑問系かどうか",
            "title": "Is Interrogative",
            "type": "boolean"
          },
          "moras": {
            "description": "モーラのリスト",
            "items": {
              "$ref": "#/components/schemas/Mora"
            },
            "title": "Moras",
            "type": "array"
          },
          "pause_mora": {
            "$ref": "#/components/schemas/Mora",
            "description": "アクセント句の末尾につく無音モーラ。null の場合は無音モーラを付けない。",
            "title": "Pause Mora"
          }
        },
        "required": [
          "moras",
          "accent"
        ],
        "title": "AccentPhrase",
        "type": "object"
      },
      "AudioQuery": {
        "description": "音声合成用のクエリ。",
        "properties": {
          "accent_phrases": {
            "description": "アクセント句のリスト",
            "items": {
              "$ref": "#/components/schemas/AccentPhrase"
            },
            "title": "Accent Phrases",
            "type": "array"
          },
          "intonationScale": {
            "description": "全体の抑揚",
            "title": "Intonationscale",
            "type": "number"
          },
          "kana": {
            "description": "[読み取り専用]AquesTalk 風記法によるテキスト。音声合成用のクエリとしては無視される",
            "title": "Kana",
            "type": "string"
          },
          "outputSamplingRate": {
            "description": "音声データの出力サンプリングレート",
            "exclusiveMinimum": 0.0,
            "title": "Outputsamplingrate",
            "type": "integer"
          },
          "outputStereo": {
            "description": "音声データをステレオ出力するか否か",
            "title": "Outputstereo",
            "type": "boolean"
          },
          "pauseLength": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "句読点などの無音時間。nullのときは無視される。デフォルト値はnull",
            "title": "Pauselength"
          },
          "pauseLengthScale": {
            "default": 1,
            "description": "句読点などの無音時間（倍率）。デフォルト値は1",
            "minimum": 0.0,
            "title": "Pauselengthscale",
            "type": "number"
          },
          "pitchScale": {
            "description": "全体の音高",
            "title": "Pitchscale",
            "type": "number"
          },
          "postPhonemeLength": {
            "description": "音声の後の無音時間",
            "minimum": 0.0,
            "title": "Postphonemelength",
            "type": "number"
          },
          "prePhonemeLength": {
            "description": "音声の前の無音時間",
            "minimum": 0.0,
            "title": "Prephonemelength",
            "type": "number"
          },
          "speedScale": {
            "description": "全体の話速",
            "exclusiveMinimum": 0.0,
            "title": "Speedscale",
            "type": "number"
          },
          "volumeScale": {
            "description": "全体の音量",
            "minimum": 0.0,
            "title": "Volumescale",
            "type": "number"
          }
        },
        "required": [
          "accent_phrases",
          "speedScale",
          "pitchScale",
          "intonationScale",
          "volumeScale",
          "prePhonemeLength",
          "postPhonemeLength",
          "outputSamplingRate",
          "outputStereo"
        ],
        "title": "AudioQuery",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Mora": {
        "description": "モーラ（子音＋母音）ごとの情報。",
        "properties": {
          "consonant": {
            "description": "子音の音素",
            "title": "Consonant",
            "type": "string"
          },
          "consonant_length": {
            "description": "子音の長さ",
            "minimum": 0.0,
            "title": "Consonant Length",
            "type": "number"
          },
          "pitch": {
            "description": "音高",
            "title": "Pitch",
            "type": "number"
          },
          "text": {
            "description": "文字",
            "title": "Text",
            "type": "string"
          },
          "vowel": {
            "description": "母音の音素",
            "title": "Vowel",
            "type": "string"
          },
          "vowel_length": {
            "description": "母音の長さ",
            "minimum": 0.0,
            "title": "Vowel Length",
            "type": "number"
          }
        },
        "required": [
          "text",
          "vowel",
          "vowel_length",
          "pitch"
        ],
        "title": "Mora",
        "type": "object"
      },
      "MorphableTargetInfo": {
        "description": "モーフィング相手としての情報。",
        "properties": {
          "is_morphable": {
            "description": "指定したキャラクターに対してモーフィングの可否",
            "title": "Is Morphable",
            "type": "boolean"
          }
        },
        "required": [
          "is_morphable"
        ],
        "title": "MorphableTargetInfo",
        "type": "object"
      },
...
      "ResourceFormat": {
        "enum": [
          "base64",
          "url"
        ],
        "type": "string"
      },
...
      "Speaker": {
        "description": "キャラクター情報",
        "properties": {
          "name": {
            "description": "名前",
            "title": "Name",
            "type": "string"
          },
          "speaker_uuid": {
            "description": "キャラクターのUUID",
            "title": "Speaker Uuid",
            "type": "string"
          },
          "styles": {
            "description": "スタイルの一覧",
            "items": {
              "$ref": "#/components/schemas/SpeakerStyle"
            },
            "title": "Styles",
            "type": "array"
          },
          "supported_features": {
            "$ref": "#/components/schemas/SpeakerSupportedFeatures",
            "description": "キャラクターの対応機能"
          },
          "version": {
            "description": "キャラクターのバージョン",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "speaker_uuid",
          "styles",
          "version"
        ],
        "title": "Speaker",
        "type": "object"
      },
      "SpeakerInfo": {
        "description": "キャラクターの追加情報",
        "properties": {
          "policy": {
            "description": "policy.md",
            "title": "Policy",
            "type": "string"
          },
          "portrait": {
            "description": "立ち絵画像をbase64エンコードしたもの、あるいはURL",
            "title": "Portrait",
            "type": "string"
          },
          "style_infos": {
            "description": "スタイルの追加情報",
            "items": {
              "$ref": "#/components/schemas/StyleInfo"
            },
            "title": "Style Infos",
            "type": "array"
          }
        },
        "required": [
          "policy",
          "portrait",
          "style_infos"
        ],
        "title": "SpeakerInfo",
        "type": "object"
      },
      "SpeakerStyle": {
        "description": "キャラクターのスタイル情報",
        "properties": {
          "id": {
            "description": "スタイルID",
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "description": "スタイル名",
            "title": "Name",
            "type": "string"
          },
          "type": {
            "default": "talk",
            "description": "スタイルの種類。talk:音声合成クエリの作成と音声合成が可能。singing_teacher:歌唱音声合成用のクエリの作成が可能。frame_decode:歌唱音声合成が可能。sing:歌唱音声合成用のクエリの作成と歌唱音声合成が可能。streaming_talk:音声合成クエリの作成と音声合成とストリーミング音声合成が可能。",
            "enum": [
              "talk",
              "singing_teacher",
              "frame_decode",
              "sing",
              "streaming_talk"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "SpeakerStyle",
        "type": "object"
      },
      "SpeakerSupportedFeatures": {
        "description": "キャラクターの対応機能の情報",
        "properties": {
          "permitted_synthesis_morphing": {
            "default": "ALL",
            "description": "モーフィング機能への対応。'ALL' は「全て許可」、'SELF_ONLY' は「同じキャラクター内でのみ許可」、'NOTHING' は「全て禁止」",
            "enum": [
              "ALL",
              "SELF_ONLY",
              "NOTHING"
            ],
            "title": "Permitted Synthesis Morphing",
            "type": "string"
          }
        },
        "title": "SpeakerSupportedFeatures",
        "type": "object"
      },
      "StyleInfo": {
        "description": "スタイルの追加情報",
        "properties": {
          "icon": {
            "description": "このスタイルのアイコンをbase64エンコードしたもの、あるいはURL",
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "description": "スタイルID",
            "title": "Id",
            "type": "integer"
          },
          "portrait": {
            "description": "このスタイルの立ち絵画像をbase64エンコードしたもの、あるいはURL",
            "title": "Portrait",
            "type": "string"
          },
          "voice_samples": {
            "description": "サンプル音声をbase64エンコードしたもの、あるいはURL",
            "items": {
              "type": "string"
            },
            "title": "Voice Samples",
            "type": "array"
          }
        },
        "required": [
          "id",
          "icon",
          "voice_samples"
        ],
        "title": "StyleInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "VOICEVOX OSS の音声合成エンジンです。",
    "title": "VOICEVOX ENGINE OSS",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/morphable_targets": {
      "post": {
        "description": "指定されたベーススタイルに対してエンジン内の各キャラクターがモーフィング機能を利用可能か返します。\n\nモーフィングの許可/禁止は`/speakers`の`speaker.supported_features.synthesis_morphing`に記載されています。\nプロパティが存在しない場合は、モーフィングが許可されているとみなします。\n返り値のスタイルIDはstring型なので注意。",
        "operationId": "morphable_targets",
        "parameters": [
          {
            "description": "コアのバージョン",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "コアのバージョン",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "integer"
                },
                "title": "Base Style Ids",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "$ref": "#/components/schemas/MorphableTargetInfo"
                    },
                    "type": "object"
                  },
                  "title": "Response Morphable Targets",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定したスタイルに対してエンジン内のキャラクターがモーフィングが可能か判定する",
        "tags": [
          "音声合成"
        ]
      }
    },
...
    "/singer_info": {
      "get": {
        "description": "UUID で指定された歌えるキャラクターの情報を返します。\n\n画像や音声はresource_formatで指定した形式で返されます。",
        "operationId": "singer_info",
        "parameters": [
          {
            "description": "キャラクターのUUID",
            "in": "query",
            "name": "speaker_uuid",
            "required": true,
            "schema": {
              "description": "キャラクターのUUID",
              "title": "Speaker Uuid",
              "type": "string"
            }
          },
          {
            "description": "画像や音声の返却形式",
            "in": "query",
            "name": "resource_format",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ResourceFormat",
              "default": "base64",
              "description": "画像や音声の返却形式"
            }
          },
          {
            "description": "コアのバージョン",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "コアのバージョン",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Singer Info",
        "tags": [
          "その他"
        ]
      }
    },
    "/singers": {
      "get": {
        "description": "歌えるキャラクターの情報の一覧を返します。",
        "operationId": "singers",
        "parameters": [
          {
            "description": "コアのバージョン",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "コアのバージョン",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Singers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Singers",
        "tags": [
          "その他"
        ]
      }
    },
    "/speaker_info": {
      "get": {
        "description": "UUID で指定された喋れるキャラクターの情報を返します。\n\n画像や音声はresource_formatで指定した形式で返されます。",
        "operationId": "speaker_info",
        "parameters": [
          {
            "description": "キャラクターのUUID",
            "in": "query",
            "name": "speaker_uuid",
            "required": true,
            "schema": {
              "description": "キャラクターのUUID",
              "title": "Speaker Uuid",
              "type": "string"
            }
          },
          {
            "description": "画像や音声の返却形式",
            "in": "query",
            "name": "resource_format",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ResourceFormat",
              "default": "base64",
              "description": "画像や音声の返却形式"
            }
          },
          {
            "description": "コアのバージョン",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "コアのバージョン",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Speaker Info",
        "tags": [
          "その他"
        ]
      }
    },
    "/speakers": {
      "get": {
        "description": "喋れるキャラクターの情報の一覧を返します。",
        "operationId": "speakers",
        "parameters": [
          {
            "description": "コアのバージョン",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "コアのバージョン",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Speakers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Speakers",
        "tags": [
          "その他"
        ]
      }
    },
...
    "/synthesis_morphing": {
      "post": {
        "description": "指定された2種類のスタイルで音声を合成、指定した割合でモーフィングした音声を得ます。\n\nモーフィングの割合は`morph_rate`で指定でき、0.0でベースのスタイル、1.0でターゲットのスタイルに近づきます。",
        "operationId": "_synthesis_morphing",
        "parameters": [
          {
            "description": "モーフィングのベースのスタイルID",
            "in": "query",
            "name": "base_speaker",
            "required": true,
            "schema": {
              "description": "モーフィングのベースのスタイルID",
              "title": "Base Speaker",
              "type": "integer"
            }
          },
          {
            "description": "モーフィングのターゲットのスタイルID",
            "in": "query",
            "name": "target_speaker",
            "required": true,
            "schema": {
              "description": "モーフィングのターゲットのスタイルID",
              "title": "Target Speaker",
              "type": "integer"
            }
          },
          {
            "description": "モーフィングの割合（0.0でベース、1.0でターゲットのスタイルに近づく）",
            "in": "query",
            "name": "morph_rate",
            "required": true,
            "schema": {
              "description": "モーフィングの割合（0.0でベース、1.0でターゲットのスタイルに近づく）",
              "maximum": 1.0,
              "minimum": 0.0,
              "title": "Morph Rate",
              "type": "number"
            }
          },
          {
            "description": "疑問系のテキストが与えられたら語尾を自動調整する",
            "in": "query",
            "name": "enable_interrogative_upspeak",
            "required": false,
            "schema": {
              "default": true,
              "description": "疑問系のテキストが与えられたら語尾を自動調整する",
              "title": "Enable Interrogative Upspeak",
              "type": "boolean"
            }
          },
          {
            "description": "コアのバージョン",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "コアのバージョン",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "2種類のスタイルでモーフィングした音声を合成する",
        "tags": [
          "音声合成"
        ]
      }
    },
...
  }
}